Fixed bug where enemy NPC turret-style weapons would not  track player
Weapon tracking uses a transform to trace targets and rotate the weapon to find a firing solution. If the transform is null, the weapon will act as if "mounted" (will only face in the direction it is installed). Solution was to default the tracking target to a "Player" object when it is null. If NPCs need to be able to find targets that are not the player, this workaround will need to be modified.